+++ code/src/cljs/sixsq/nuvla/ui/apps_application/spec.cljs
@@ -1,6 +1,7 @@
 (ns sixsq.nuvla.ui.apps-application.spec
   (:require-macros [sixsq.nuvla.ui.utils.spec :refer [only-keys]])
   (:require [clojure.spec.alpha :as s]
+            [sixsq.nuvla.ui.utils.general :as general-utils]
             [sixsq.nuvla.ui.utils.spec :as spec-utils]
             [taoensso.timbre :as log]))
 
@@ -10,7 +11,8 @@
 
 ; Docker-compose
 
-(s/def ::docker-compose spec-utils/nonblank-string)
+(s/def ::docker-compose (s/and spec-utils/nonblank-string
+                               #(first (general-utils/check-yaml %))))
 
 
 ; Files